Made the time code more sane and changed some types and their meaning around.
Time is now measured in milliseconds, instead of the old format which used the upper 32bit for the seconds the lower 32bit for the milliseconds.
There is now also a type for holding unix timestamps (in seconds) as well as convenient functions to convert the two time types around.

There is also code from the FreeBSD project which is used for 64bit integer division (see /sys/libc/bsd/README.md)

+++ libkernel/libio/IOThread.cpp
@@ -36,7 +36,7 @@ extern "C"
 	thread_t __io_threadCreate(void *entry, void *owner, void *arg);
 	thread_t __io_threadID();
 	void __io_threadSetName(thread_t id, const char *name);
-	void __io_threadSleep(thread_t id, uint32_t time);
+	void __io_threadSleep(thread_t id, uint64_t time);
 	void __io_threadWakeup(thread_t id);
 	void sd_yield();
 }
@@ -132,7 +132,7 @@ void IOThread::threadEntry()
 	release();
 }
 
-void IOThread::sleep(uint32_t time)
+void IOThread::sleep(uint64_t time)
 {
 	__io_threadSleep(_id, time);
 }
